+++ src/PSRule/Pipeline/Runs/IRun.cs
@@ -1,8 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using PSRule.Data;
 using PSRule.Definitions;
 using PSRule.Runtime;
+using PSRule.Runtime.Binding;
 
 namespace PSRule.Pipeline.Runs;
 
@@ -36,4 +38,9 @@ public interface IRun : IConfiguration
     /// A list of rules that are part of the run.
     /// </summary>
     IRuleGraph Rules { get; }
+
+    /// <summary>
+    /// Evaluate bound properties for a target object.
+    /// </summary>
+    ITargetBindingResult Bind(ITargetObject targetObject);
 }