+++ docs/changelog.md
@@ -45,6 +45,8 @@ What's changed since pre-release v3.0.0-B0453:
     [#2891](https://github.com/microsoft/PSRule/issues/2891)
   - **Important change**: Remove pass fail streams by @BernieWhite.
     [#2892](https://github.com/microsoft/PSRule/issues/2892)
+  - **Important change**: Cleanup PreferTargetInfo option by @BernieWhite.
+    [#2994](https://github.com/microsoft/PSRule/issues/2994)
   - Added GitHub Actions support to CLI by @BernieWhite.
     [#2824](https://github.com/microsoft/PSRule/issues/2824)
   - Added Azure Pipelines support to CLI by @BernieWhite.